requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bidict==0.22.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SocketIO==5.3.4
frozenlist==1.3.3
gitignore-parser==0.1.3
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.3
openai==0.27.7
python-dotenv==1.0.0
python-engineio==4.4.1
python-socketio==5.8.0
PyYAML==6.0
redis==4.5.5
regex==2023.5.5
requests==2.31.0
six==1.16.0
tiktoken==0.4.0
tqdm==4.65.0
urllib3==2.0.2
Werkzeug==2.3.4
yarl==1.9.2
